Preliminary MIDI parser working

+++ src/main/java/Note.java
@@ -2,10 +2,10 @@
  * Created by jack on 3/22/2016.
  */
 public class Note {
-    private double duration;
-    private int pitch;
+    private double duration; // in beats
+    private int pitch; // in MIDI keyboard
 
-    public Note(float duration, int pitch) {
+    public Note(double duration, int pitch) {
         this.duration = duration;
         this.pitch = pitch;
     }
